# apps/control-api/src/app/operations/operations.controller.ts
/**
 * REST surface for asynchronous operation readback — the polling endpoint.
 *
 * A client that receives `202 Accepted` from a mutation polls `GET :operationId` here until
 * the operation reaches a terminal state. Responses come from the read projection, so they
 * reflect the last workflow event the projection worker has applied.
 *
 * Transport adapter only. `OperationGrpcController` is the gRPC equivalent.
 *
 * @see docs/architecture/state-machines.md
 */
import {
  Controller,
  DefaultValuePipe,
  Get,
  Inject,
  Param,
  ParseIntPipe,
  ParseUUIDPipe,
  Query,
  UseGuards,
} from '@nestjs/common';
import { ControlPlaneApplication, type Actor } from '@private-cloud/application';
import { CurrentActor } from '../auth/actor.decorator';
import { OidcAuthGuard } from '../auth/oidc-auth.guard';
import { CONTROL_PLANE_APPLICATION } from '../tokens';

@Controller('v1/projects/:projectId/operations')
@UseGuards(OidcAuthGuard)
export class OperationsController {
  public constructor(
    @Inject(CONTROL_PLANE_APPLICATION) private readonly application: ControlPlaneApplication,
  ) {}

  @Get()
  public list(
    @CurrentActor() actor: Actor,
    @Param('projectId', new ParseUUIDPipe()) projectId: string,
    @Query('limit', new DefaultValuePipe(50), ParseIntPipe) limit: number,
    @Query('cursor') cursor?: string,
  ) {
    return this.application.listOperations(actor, projectId, limit, cursor);
  }

  @Get(':operationId')
  public get(
    @CurrentActor() actor: Actor,
    @Param('projectId', new ParseUUIDPipe()) projectId: string,
    @Param('operationId', new ParseUUIDPipe()) operationId: string,
  ) {
    return this.application.getOperation(actor, projectId, operationId);
  }
}