# packages/contracts/openapi/control-plane.v1.yaml
openapi: 3.1.0
info:
  title: Private Cloud Control Plane API
  version: 1.0.0
  summary: Provider-neutral virtual-machine lifecycle control plane
  description: |
    Accepts durable asynchronous lifecycle commands and exposes project-scoped read models.
    A `202` response means intent and its operation record committed; it does not mean provider work completed.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://cloud.example.invalid
    description: Environment-specific Gateway API endpoint
security:
  - bearerAuth: []
...
paths:
...
  /v1/projects/{projectId}/operations:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
      - $ref: './components.v1.yaml#/components/parameters/Cursor'
      - $ref: './components.v1.yaml#/components/parameters/Limit'
    get:
      operationId: listOperations
      tags: [Operations]
      summary: List project operations
      x-audience: tenant
      x-idempotency: read-only
      x-error-codes: [VALIDATION_FAILED, PROJECT_ACCESS_DENIED]
      x-requirements: [SRS-FR-002, SRS-FR-021, SRS-FR-022]
      parameters:
        - name: instanceId
          in: query
          schema: { type: string, format: uuid }
        - name: state
          in: query
          schema:
            $ref: './components.v1.yaml#/components/schemas/OperationState'
      responses:
        '200':
          description: Operation page
          content:
            application/json:
              schema:
                $ref: './components.v1.yaml#/components/schemas/OperationPage'
        '400': { $ref: './components.v1.yaml#/components/responses/BadRequest' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
  /v1/projects/{projectId}/operations/{operationId}:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/ProjectId'
      - $ref: './components.v1.yaml#/components/parameters/OperationId'
    get:
      operationId: getOperation
      tags: [Operations]
      summary: Read one project operation
      x-audience: tenant
      x-idempotency: read-only
      x-error-codes: [PROJECT_ACCESS_DENIED, OPERATION_NOT_FOUND]
      x-requirements: [SRS-FR-002, SRS-FR-021, SRS-FR-022]
      responses:
        '200':
          description: Operation detail
          content:
            application/json:
              schema:
                $ref: './components.v1.yaml#/components/schemas/Operation'
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
...
  /v1/admin/operations/{operationId}:
    parameters:
      - $ref: './components.v1.yaml#/components/parameters/OperationId'
    get:
      operationId: getAdministrativeOperation
      tags: [Administration]
      summary: Read one operation with recovery metadata
      x-audience: administrator
      x-idempotency: read-only
      x-error-codes: [ADMIN_REQUIRED, OPERATION_NOT_FOUND]
      x-requirements: [SRS-FR-004, SRS-FR-023]
      responses:
        '200':
          description: Administrative operation detail
          content:
            application/json:
              schema: { $ref: './components.v1.yaml#/components/schemas/AdministrativeOperation' }
        '401': { $ref: './components.v1.yaml#/components/responses/Unauthorized' }
        '403': { $ref: './components.v1.yaml#/components/responses/Forbidden' }
        '404': { $ref: './components.v1.yaml#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT